Migrate TVRename to .NET 8 / C# 12 and fix build regressions

+++ TVRename/Sources/TMDB/LocalCache.cs
@@ -780,7 +780,7 @@ public class LocalCache : MediaCache, iMovieSource, iTVSource
                     Name = downloadedEpisode.Name,
                     Overview = downloadedEpisode.Overview,
                     FirstAired = downloadedEpisode.AirDate,
-                    AiredEpNum = downloadedEpisode.EpisodeNumber,
+                    AiredEpNum = (int)downloadedEpisode.EpisodeNumber,
                     AiredSeasonNumber = downloadedEpisode.SeasonNumber,
                     ProductionCode = downloadedEpisode.ProductionCode,
                     EpisodeId = downloadedEpisode.Id,
